refactor(#726): move the track-release kernel + song-as-track + swapped strategies into lookup/strategies/
Pure code motion, PR 4a of the orchestrator decomposition (#722). The shared release-to-library kernel (_match_track_releases_to_library), search_album_fuzzy, and SONG_AS_TRACK_CONFIDENCE (sole consumer is the kernel, per the manifest comment on #726) move verbatim into the new lookup/strategies/track_release_matching.py; search_song_as_track lands in the existing song_as_track.py shell and _narrow_swapped_by_track + search_with_alternative_interpretation in the existing swapped_interpretation.py shell. The orchestrator's build_strategies(...) wiring now imports the two execute functions from their strategy modules, and search_compilations_for_track / find_library_albums_with_cached_track (which stay until PR 4b / PR 5) import search_album_fuzzy from the kernel module. No re-exports; the strategies package still has no dependency on lookup.orchestrator, so the existing import direction (orchestrator -> strategies) is preserved. Only additions are the module docstring and logger lines the program mechanics allow.
